+++ Core/Architecture/Infrastructure/Runtime/Initialization/InitializableComponent.cs
@@ -1,0 +1,17 @@
+using UnityEngine;
+
+public abstract class InitializableComponent : MonoBehaviour, IInitializable
+{
+    protected bool isInitialized;
+
+    public void Initialize()
+    {
+        if (isInitialized)
+            return;
+
+        isInitialized = true;
+        OnInitialized();
+    }
+
+    protected abstract void OnInitialized();
+}